Fixes Issue #3
Fixes Issue #3
This commit fixes issue #3 by checking for the file lock on the source and destination files.  If a lock is found, it asks the user what they want to do.  After it is determined that no locks exist, it attempts a file copy.  If the file copy fails, it renames the file to append _NEW onto the name and tries the copy a second time.  If it fails the second time, it gives up but lets the user know to put in an issue on github and what information to provide.
With my quick testing, it appears to work when the destination file is locked to rename it.
The renamed version of the file is passed back to the calling function along with a success value.

This should correct issue #3 or at least give end users a chance to fix things on their own (delete the locked file manually).
Also fixed a typo in the lock checking loop.  Not entirely sure how it wasn't stuck in an infinite loop previously when checking for locks.